requirements.txt
fastapi
uvicorn[standard]
weasyprint
markdown
jinja2
python-multipart
Pillow
language-tool-python
google-generativeai
google-genai
python-dotenv
EbookLib
qrcode
streamlit
edge-tts
wikipedia-api
bs4
